DF-8850 eliminated some warnings;
Former-commit-id: 281e38ba80f71eec31e9b6a8cbb3a6f00960585e

+++ Sources/Internal/Render/RHI/GLES2/rhi_CommandBufferGLES2.cpp
@@ -823,7 +823,7 @@ void CommandBufferGLES2_t::Execute()
         case GLES2__SET_VERTEX_DATA:
         {
             Handle vb = (Handle)(arg[0]);
-            unsigned stream_i = arg[1];
+            unsigned stream_i = uint32(arg[1]);
 
             if (cur_vb[stream_i] != vb)
             {
